make sure value exists and that it is array before checking for keys

+++ includes/Customer.php
@@ -39,7 +39,9 @@ class Customer {
     public static function collect() {
         $bh_cdata = Transient::get( self::TRANSIENT );
 
-        if ( ! array_key_exists( 'signup_date', $bh_cdata ) ||
+        if ( $bh_data &&
+             is_array( $bh_cdata ) &&
+             ! array_key_exists( 'signup_date', $bh_cdata ) ||
              ! array_key_exists( 'plan_subtype', $bh_cdata ) ) {
             $bh_cdata = false;
             Transient::delete( self::TRANSIENT );
@@ -48,7 +50,7 @@ class Customer {
         if ( ! $bh_cdata ) {
             $guapi    = self::get_account_info();
             $mole     = self::get_onboarding_info();
-            if ( empty( $guapi ) || empty( $mole ) ) {
+            if ( empty( $guapi ) ) {
                 return false;
             }
             $bh_cdata = array_merge( $guapi, array( 'meta' => $mole ) );
